Move matches test in matches module
